current new state
